testovoe/requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
python-decouple~=3.8
djangorestframework-xml~=2.0.0